Make off time exclusive
When a user says "Turn lights on at 8PM, off at 10PM", they usually mean that at 8:00:00 PM the lights turn on, and at 10:00:00 PM the lights turn off. Hence, the range is meant inclusively of the on time, and exclusively of the off time. This patch implements this behavior.